Fix the ZlibStreamInflator
`Ionic.Zlib.DeflateStream` works very differently to
`System.IO.Compression.DeflateStream`.
The former wraps around an output stream and can be written to
repeatedly, reach the inflated output incrementally.
The latter wraps an input stream and can be read from repeatedly.
The pixie messages received by the client form part of a continuous
deflate stream and must be inflated within the context of previously
inflated messages.
To that end the functionality of the `DeflateStream` must be inverted in
order to behave more like the Ionic version.